Using observables
There are many ways to get streams of bytes.  The readable stream API for
Node is crazy complicated and brittle.  As such, I'm going to convert the
parser so that it works strictly off an observable of buffers.  This means
I can write it once for pretty much any potential should (streams, files,
network traffic, whatever).  Then I can implement functions that return
observables for various sources (and test those functions in the absence
of the parser, to simplify testing).