Added flag to extract part data
There was a nasty issue with having the
file to read open twice, but I resolved that.
I also got rid of the whole regexp stuff and
just put in predicates.  This is completely
general and should all anything in the
future.  I'd still like to have "glob" pattern
support, but minimatch had too many
issues.